Add support for css modules; upgrade dependencies
