+++ .eslintrc
@@ -30,7 +30,6 @@
     "node": true
   },
   "globals": {
-    '__CORDOVA__': true,
     '__BASE_URL__': true,
     expect: true
   },
