+++ common/css/vendors.scss
@@ -1,2 +1,2 @@
 // vendor styles
-@import 'vendors/bootstrap';
+@import 'vendors/include-media';
