--- common/css/vendors/_bootstrap.scss
@@ -1,2 +1,0 @@
-$icon-font-path: "~bootstrap-sass/assets/fonts/bootstrap/";
-@import "~bootstrap-sass/assets/stylesheets/bootstrap";
