+++ common/css/vendors/_include-media.scss
@@ -1,0 +1,1 @@
+@import "~include-media/dist/include-media";
