+++ common/js/components/Header.js
@@ -1,11 +1,12 @@
 import React, { Component } from 'react';
 import { Link } from 'react-router';
+import css from './Header.scss';
 
 export default class Header extends Component {
   render() {
     return (
       <header>
-        <ul>
+        <ul className={css.headerLinks}>
           <li><Link to="/">Home</Link></li>
           <li><Link to="/example">Example</Link></li>
           <li><Link to="/example/2">Example 2</Link></li>
