+++ common/js/components/Header.scss
@@ -1,0 +1,4 @@
+
+.headerLinks {
+  color: red;
+}
