+++ common/layouts/server.html
@@ -10,6 +10,7 @@
 <body>
   <div id="app"><%= html %></div>
   <script>window.__INITIAL_STATE__ = <%= JSON.stringify(initialState) %>;</script>
-  <script src="<%= bundleJs %>"></script>
+  <script src="<%= vendorJs %>"></script>
+  <script src="<%= appJs %>"></script>
 </body>
 </html>
