+++ nodemon.json
@@ -2,6 +2,7 @@
   "watch": [
     "server/**/*.js",
     "server/**/*.json",
+    "common/js/**/*.js",
     "common/layouts/server.html"
   ]
 }
