+++ package.json
@@ -4,6 +4,9 @@
   "version": "1.0.0",
   "license": "MIT",
   "main": "client/index.js",
+  "config": {
+    "css": "[name]__[local]__[hash:base64:5]"
+  },
   "repository": {
     "type": "git",
     "url": "https://github.com/combine/react-redux-starter.git"
@@ -16,9 +19,9 @@
     "start": "npm run dev:build && npm run dev:start",
     "dev:start": "npm run dev:start:client & npm run dev:start:server",
     "dev:start:server": "NODE_ENV=development $(npm bin)/nodemon server",
-    "dev:start:client": "NODE_ENV=development node webpack/development.server",
-    "dev:build": "NODE_ENV=development $(npm bin)/webpack --progress --colors --display-error-details --config webpack/development.js",
-    "prod:build": "NODE_ENV=production $(npm bin)/webpack -p --optimize-dedupe --config webpack/production.js",
+    "dev:start:client": "NODE_ENV=development node webpack/dev.hot",
+    "dev:build": "NODE_ENV=development $(npm bin)/webpack --progress --colors --display-error-details --config webpack/dev.js",
+    "prod:build": "NODE_ENV=production $(npm bin)/webpack -p --optimize-dedupe --config webpack/prod.js",
     "prod:start": "NODE_ENV=production pm2 start server --name='callai'",
     "test": "$(npm bin)/mocha --recursive --compilers js:babel-register",
     "test:all": "$(npm bin)/mocha --recursive --compilers js:babel-register $(find . -path ./node_modules -prune -o -name '*.test.js' -print)",
@@ -28,27 +31,9 @@
     "postinstall": "if [ \"$NODE_ENV\" = \"production\" ]; then npm run prod:build; fi"
   },
   "devDependencies": {
-    "babel-eslint": "6.1.2",
-    "chai": "3.5.0",
-    "chai-immutable": "1.6.0",
-    "eslint": "3.4.0",
-    "eslint-plugin-react": "6.2.0",
-    "mocha": "3.0.2",
-    "nodemon": "1.10.2",
-    "react-transform-catch-errors": "1.0.2",
-    "react-transform-hmr": "1.0.4",
-    "redbox-react": "1.3.0",
-    "redux-devtools": "3.3.1",
-    "redux-devtools-dock-monitor": "1.1.1",
-    "redux-devtools-log-monitor": "1.0.11",
-    "webpack-dev-server": "1.15.0",
-    "yargs": "5.0.0"
-  },
-  "dependencies": {
-    "app-module-path": "1.1.0",
-    "autoprefixer": "6.4.0",
     "babel": "6.5.2",
     "babel-core": "6.14.0",
+    "babel-eslint": "6.1.2",
     "babel-loader": "6.2.5",
     "babel-plugin-react-transform": "2.0.2",
     "babel-plugin-resolver": "1.0.0",
@@ -58,29 +43,52 @@
     "babel-plugin-transform-class-properties": "6.11.5",
     "babel-plugin-transform-decorators": "6.13.0",
     "babel-plugin-transform-object-rest-spread": "6.8.0",
-    "babel-polyfill": "6.13.0",
     "babel-preset-es2015": "6.14.0",
     "babel-preset-react": "6.11.1",
     "babel-register": "6.14.0",
-    "bootstrap-sass": "3.3.7",
-    "classnames": "2.2.5",
-    "compression": "1.6.2",
+    "chai": "3.5.0",
+    "chai-immutable": "1.6.0",
     "compression-webpack-plugin": "0.3.1",
     "css-loader": "0.24.0",
+    "eslint": "3.4.0",
     "eslint-loader": "1.5.0",
+    "eslint-plugin-react": "6.2.0",
     "expose-loader": "0.7.1",
-    "express": "4.14.0",
     "extract-text-webpack-plugin": "1.0.1",
     "file-loader": "0.9.0",
+    "html-webpack-plugin": "2.22.0",
+    "json-loader": "0.5.4",
+    "mocha": "3.0.2",
+    "nodemon": "1.10.2",
+    "postcss-loader": "0.11.0",
+    "react-transform-catch-errors": "1.0.2",
+    "react-transform-hmr": "1.0.4",
+    "redbox-react": "1.3.0",
+    "redux-devtools": "3.3.1",
+    "redux-devtools-dock-monitor": "1.1.1",
+    "redux-devtools-log-monitor": "1.0.11",
+    "sass-loader": "4.0.0",
+    "style-loader": "0.13.1",
+    "url-loader": "0.5.7",
+    "webpack": "1.13.2",
+    "webpack-dev-server": "1.15.0",
+    "webpack-sources": "0.1.2",
+    "yargs": "5.0.0"
+  },
+  "dependencies": {
+    "app-module-path": "1.1.0",
+    "autoprefixer": "6.4.0",
+    "babel-polyfill": "6.13.0",
+    "classnames": "2.2.5",
+    "compression": "1.6.2",
+    "css-modules-require-hook": "4.0.2",
+    "express": "4.14.0",
     "font-awesome": "4.6.3",
     "history": "3.0.0",
-    "html-webpack-plugin": "2.22.0",
     "include-media": "1.4.8",
     "isomorphic-fetch": "2.2.1",
-    "json-loader": "0.5.4",
     "lodash": "4.15.0",
     "node-sass": "3.8.0",
-    "postcss-loader": "0.11.0",
     "react": "15.3.1",
     "react-dom": "15.3.1",
     "react-redux": "4.4.5",
@@ -90,12 +98,8 @@
     "redux-localstorage": "0.4.1",
     "redux-thunk": "2.1.0",
     "reselect": "2.5.3",
-    "sass-loader": "4.0.0",
+    "resolve-url-loader": "1.6.0",
     "serve-static": "1.11.1",
-    "style-loader": "0.13.1",
-    "url-loader": "0.5.7",
-    "webpack": "1.13.2",
-    "webpack-isomorphic-tools": "2.5.8",
-    "webpack-sources": "0.1.2"
+    "webpack-isomorphic-tools": "2.5.8"
   }
 }
