--- scripts/cordova/config/fabric.json.example
@@ -1,4 +1,0 @@
-{
-  "apiKey": "my-api-key",
-  "apiSecret": "my-build-secret"
-}
