--- scripts/cordova/config/release.json.example
@@ -1,11 +1,0 @@
-{
-  "android": {
-    "release": {
-      "keystore": "path/to/keystore.jks",
-      "storePassword": "myStorePassword",
-      "alias": "myKeyAlias",
-      "password" : "myKeyPassword",
-      "keystoreType": "jks"
-    }
-  }
-}
