--- scripts/cordova/index.js
@@ -1,1 +1,0 @@
-require('./build')(require('yargs').argv);
